+++ Data/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -335,8 +335,8 @@ namespace OnlineMobileRecharge.Data.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Package_Id"));
 
-                    b.Property<int>("Package_Data")
-                        .HasColumnType("int");
+                    b.Property<long>("Package_Data")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("Package_Description")
                         .IsRequired()
@@ -349,17 +349,17 @@ namespace OnlineMobileRecharge.Data.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("Package_Off_Net_Mins")
-                        .HasColumnType("int");
+                    b.Property<long>("Package_Off_Net_Mins")
+                        .HasColumnType("bigint");
 
-                    b.Property<int>("Package_On_Net_Mins")
-                        .HasColumnType("int");
+                    b.Property<long>("Package_On_Net_Mins")
+                        .HasColumnType("bigint");
 
-                    b.Property<int>("Package_Price")
-                        .HasColumnType("int");
+                    b.Property<double>("Package_Price")
+                        .HasColumnType("float");
 
-                    b.Property<int>("Package_SMS")
-                        .HasColumnType("int");
+                    b.Property<long>("Package_SMS")
+                        .HasColumnType("bigint");
 
                     b.Property<int>("Package_Type")
                         .HasColumnType("int");
